Abstract API fetching methods into parent class
